Crud movies

+++ src/common/file.cleanup-interceptor.ts
@@ -10,6 +10,10 @@ import * as fs from 'fs';
 
 interface RequestWithFile extends Request {
   file?: Express.Multer.File;
+  files?: {
+    image?: Express.Multer.File[];
+    video?: Express.Multer.File[];
+  };
 }
 
 @Injectable()
@@ -21,7 +25,24 @@ export class FileCleanupInterceptor implements NestInterceptor {
     return next.handle().pipe(
       catchError((error) => {
         if (request.file) {
-          fs.unlinkSync(request.file.path);
+          if (fs.existsSync(request.file.path)) {
+            fs.unlinkSync(request.file.path);
+          }
+        } else if (request.files) {
+          if (request.files.image) {
+            request.files.image.forEach((file) => {
+              if (fs.existsSync(file.path)) {
+                fs.unlinkSync(file.path);
+              }
+            });
+          }
+          if (request.files.video) {
+            request.files.video.forEach((file) => {
+              if (fs.existsSync(file.path)) {
+                fs.unlinkSync(file.path);
+              }
+            });
+          }
         }
         throw error;
       }),
